fix(vlm): address review — explicit media provenance, Nemotron helpers relocated
- Replace the key-presence reattach guard with an explicit provenance
  marker (ROLLOUT_MATCHED_MEDIA_KEY): the attach sets it only on turns it
  actually REPAIRED to the rollout's placeholder runs, and the driver-side
  static reattach skips (and consumes) exactly those turns. Unmarked
  values — including placeholder or stale payloads — are replaced as
  before, restoring the documented behavior of
  test_reattach_original_multimodal_payloads_is_media_only_and_turn_aligned,
  and unrepaired turns keep the shared-tensor restore across a prompt
  group's repeated rows.
- Move the Nemotron-specific parity logic (placeholder-run parsing, static
  budget prediction, exact-count re-processing and grid math) out of
  multimodal_utils into the existing Nemotron helper module
  (nemo_rl/environments/nemotron_utils.py). The generic attach keeps only
  the verification contract and delegates the repair via a local import;
  processors without the placeholder grammar degrade gracefully as before.
- Add a regression test: a marked turn keeps its rollout-matched tensors
  and the marker is consumed, while unmarked representations are still
  restored from the static source.
